# Optimize school check in Educator#is_authorized_for_student