--- a/app/models/educator.rb
+++ b/app/models/educator.rb
@@ -65,7 +65,7 @@ class Educator < ActiveRecord::Base
 
     return false if self.restricted_to_sped_students && !(student.program_assigned.in? ['Sp Ed', 'SEIP'])
     return false if self.restricted_to_english_language_learners && student.limited_english_proficiency == 'Fluent'
-    return false if self.school.present? && self.school != student.school
+    return false if self.school_id.present? && self.school_id != student.school_id
 
     return true if self.schoolwide_access? || self.admin? # Schoolwide admin
     return true if self.has_access_to_grade_levels? && student.grade.in?(self.grade_level_access) # Grade level access