@kevinrobinson 2017-10-03
@alexsoble @jhilde This is trying to help with https://github.com/studentinsights/studentinsights/issues/1080 and related to https://github.com/studentinsights/studentinsights/pull/1127.