@jhilde 2017-10-03
@kevinrobinson I'm trying to understand why this improves the timing so much. I would have assumed this reduces the number of queries required by ActiveRecord but it doesn't seem to.

It looks like ActiveRecord is querying for school.* way too much. I've changed the logic in is_authorized_for_student to use educator.school_id rather than educator.school and similarly for student. But there's still something causing all those queries.