@jhilde 2017-10-03
Just kidding...hate when I get too excited and forget to reload! in rails console. The third line is changed to remove lots of superfluous select * from schools where id = blah.

def is_authorized_for_student(student)
    return true if self.districtwide_access?

    return false if self.restricted_to_sped_students && !(student.program_assigned.in? ['Sp Ed', 'SEIP'])
    return false if self.restricted_to_english_language_learners && student.limited_english_proficiency == 'Fluent'
    return false if self.school_id.present? && self.school_id != student.school_id

    return true if self.schoolwide_access? || self.admin? # Schoolwide admin
    return true if self.has_access_to_grade_levels? && student.grade.in?(self.grade_level_access) # Grade level access
    return true if student.in?(self.students) # Homeroom level access
    return true if student.in?(self.section_students) # Section level access
    false
  end