@jhilde 2017-10-03
Ok, didn’t notice the file that changed. I thought the mapping was the only mod.

Yes, great minds think alike. Some are just a little slower until the coffee kicks in.