@alexsoble 2017-10-03
🔀  Merged, 🚢  shipped, and re-scheduled the data import job to run nightly — excited to check out the optimization effect on tomorrow's data import!