@kevinrobinson 2017-10-03
Cool, thanks for shipping @alexsoble and @jhilde for reviewing! 👍 

@jhilde Yeah it's surprising to me that this has such as effect.  I think in general, it's often really hard to reason about performance just by reading code.  There's caching and optimization strategies on the Ruby side with Active Record, GC on the pure Ruby side, and then there's caching and optimization strategies on the DB side with the network between them.

I started looking at this because I was surprised that this job could take so long and as a first step ran this method and saw it took a lot longer than I expected: https://github.com/studentinsights/studentinsights/pull/1118#issuecomment-333218223.  I came in assuming the heavy lifting was going to be in the homeroom or section queries towards the end.  In poking around I ran each line of the method in development to see what queries it was running, and noticed that this line pulled in the student records as you found, so I tried this change on the test case I made in the PR description.  I never would have thought it would have been a significant impact.  But measuring the change showed that it did, so when I saw that I stopped doing any other optimizations and confirmed that on a wider sample set so we could ship it.

Importantly, this might not be the main bottleneck on the nightly job either (the searchbar feature is writing a few 100kb of JSON for some rows if I remember right).  But we could use the same strategy there of making a test case and measuring it.  Use a sample of educators could be helpful there too for speeding up the iteration time when investigating or trying changes likes this.